requirements.txt
anthropic
google-generativeai
fal-client
python-dotenv
pyyaml
requests
tqdm
colorama
streamlit
